rename package folder name to all small letters
